% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complete_commut.R
\name{Complete_commut}
\alias{Complete_commut}
\title{adding commutation numbers to a mortality table}
\usage{
Complete_commut(
  lxTable,
  i,
  dim_names = NULL,
  inc_name = "age",
  messages = FALSE
)
}
\arguments{
\item{lxTable}{2 columns : age and lx}

\item{i}{actualisation rate}

\item{dim_names}{= c("sexe")}

\item{inc_name}{= "age"}

\item{messages}{= FALSE}
}
\description{
from lxTable and given act rate, adds commut columns
}
