% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_tables.R
\name{dep_table}
\alias{dep_table}
\title{Long-term care EPV}
\usage{
dep_table(
  inputs,
  detail = F,
  dim_age_dep_name = "dim_age_dep",
  i = 0,
  tmrMsg = F,
  mrgMsg = F,
  maxRows = 1e+07,
  dftResRate = 0,
  filterAgeSous = T,
  ageSousMin = 0
)
}
\arguments{
\item{inputs}{3 tables : incidence rate / mortality / dependents mortality}

\item{detail}{if FALSE : only returns EPV by age. previous name : detailedRes}

\item{dim_age_dep_name}{name of the column giving age when turning dependent}

\item{i}{actualisation rate --> not yet}

\item{tmrMsg}{bool : print messages for timer function. prev : timer_messages}

\item{mrgMsg}{bool : print tables sizes. prev : merge_messages}

\item{maxRows}{set a limit for the number of lines to avoid crashes}

\item{dftResRate}{if no t_res provided, which res rate should be used (cst)}

\item{filterAgeSous}{should the function filter on (age_vis - anc_ct)}

\item{ageSousMin}{if filterAgeSous, which age (age_vis - anc_ct) is the minimum}
}
\description{
compute Expected Present Value of long-term care annuities, in
 order to help for PRC (risk-increase provision ?)
}
