% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendLifeTable.R
\name{extendLifeTable}
\alias{extendLifeTable}
\title{extend a life Table by qx}
\usage{
extendLifeTable(lifeTbl, nRowsAdd, reOrder = TRUE)
}
\arguments{
\item{lifeTbl}{the table to extend}

\item{nRowsAdd}{number of rows to add}

\item{reOrder}{should the data be reordered by dims + inc after treatment}
}
\value{
the same columns, with more rows
}
\description{
add lines at the end of a life table, based on the last qx
}
