% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop.R
\name{generate_pop}
\alias{generate_pop}
\title{generating random population}
\usage{
generate_pop(
  tableLength = 1,
  dteVision = Sys.Date(),
  prop_F = 0.5,
  min_iD = 1,
  min_sub = ISOdate(year(Sys.Date()), 1, 1),
  max_sub = Sys.Date(),
  browse = F,
  age_min = 18,
  age_max = 30,
  messages = F,
  getTime = F
)
}
\arguments{
\item{tableLength}{number of rows}

\item{dteVision}{vision date}

\item{prop_F}{proportion of women in population, betwenn 0 and 1}

\item{min_iD}{starting iD. Will be incremented by 1 on each new row}

\item{min_sub}{minimum subscription date}

\item{max_sub}{maximum subscription date}

\item{browse}{for debugging purposes}

\item{age_min}{minimum age}

\item{age_max}{maximum age}

\item{messages}{should the function print messages while running (info only)}

\item{getTime}{if True, returns an object containing generated data and timing of different steps}
}
\description{
création ou import d'une table de pop --> sPop contenant les 5 variables sources nécessaires
}
