% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commut_vars.R
\name{get_commut_vars}
\alias{get_commut_vars}
\title{ajouter des nombres de commut a une table}
\usage{
get_commut_vars(
  popTable,
  commutTable,
  commutVars,
  cle_commut = "age",
  cle_commut_pop = cle_commut,
  cle_commut_com = cle_commut,
  key_pattern = "dim_|inc_"
)
}
\arguments{
\item{popTable}{pop table used as input}

\item{commutTable}{mortality table to get commut from}

\item{commutVars}{required commut columns}

\item{cle_commut}{how to join tables, default = by age}

\item{cle_commut_pop}{to use if key is not the same for pop and commut}

\item{cle_commut_com}{to use if key is not the same for pop and commut}
}
\description{
obtenir les nombres de commutation a partir d'une table de pop
}
