% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupIncs.R
\name{groupIncs}
\alias{groupIncs}
\title{function to group incs when a life table has 2 incs}
\usage{
groupIncs(data, incPattern = "inc_")
}
\arguments{
\item{data}{the life table with 2 incs}

\item{incPattern}{default : "inc_". Pattern to identify inc columns}
}
\description{
given a table, return the same table with a new dim and an
incgrp (and sorted)
}
