% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAll.R
\name{importAll}
\alias{importAll}
\title{fonction pour importer plusieurs fichiers et les concatener dans une meme table}
\usage{
importAll(importFunction = NULL, path = ".", pattern = "", ignore.case = FALSE)
}
\arguments{
\item{importFunction}{if you know which function you want to use}

\item{path}{argument passed to list.fles}

\item{pattern}{argument passed to list.fles}

\item{ignore.case}{argument passed to list.fles}
}
\value{
the concatenated table
}
\description{
This function imports all files selected, concatenates them in a single table and adds an fName variable
}
