% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLifeTables.R
\name{mergeLifeTables}
\alias{mergeLifeTables}
\title{merge (cross-join ?) life tables}
\usage{
mergeLifeTables(
  XTbl,
  YTbl,
  valPatt = "^qx$",
  xName,
  yName,
  maxRows = 1e+07,
  message = F
)
}
\arguments{
\item{XTbl}{"left" table}

\item{YTbl}{"right" table}

\item{valPatt}{pattern for value columns}

\item{xName}{name of X tabl, used to rename duplicate columns}

\item{yName}{name of Y tabl, used to rename duplicate columns}

\item{maxRows}{if expected rows after merge are higher than maxRows, function}

\item{message}{show expected and observed lines to check calculations
will stop executing}
}
\description{
given 2 prefixed life tables, merges them on common dims and
makes a cross-join for exclusive dims
}
