% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPRCRes.R
\name{plotPRCRes}
\alias{plotPRCRes}
\title{Describe IRP results}
\usage{
plotPRCRes(data, var = "age_vis")
}
\arguments{
\item{data}{the final table giving IRP}

\item{var}{default age_vis, the variable on x axis}
}
\value{
a plot of IRP by age (or by chosen var)
}
\description{
Quick plots describing IRP (PRC in french) results
}
