% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_to_lx.R
\name{qx_to_lx}
\alias{qx_to_lx}
\title{qx to lx}
\usage{
qx_to_lx(data, dimsNames = NULL, incName, qxName = "qx")
}
\arguments{
\item{data}{the table}

\item{dimsNames}{should be close to parameters from complete_commut function}

\item{incName}{should be close to parameters from complete_commut function}

\item{qxName}{default : "qx". if the name is different, provide it}
}
\value{
the lx column
}
\description{
given a mortality table with qx, deduce the lx
}
