% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complete_commut.R
\name{Complete_commut}
\alias{Complete_commut}
\title{adding commutation numbers to a mortality table}
\usage{
Complete_commut(
  lxTable,
  i = 0,
  dimNames = "dim_.*",
  incName = "(inc_.*)|age|x",
  valName = "(val_.*)|lx",
  messages = FALSE,
  keep_order = TRUE
)
}
\arguments{
\item{lxTable}{2 columns : age and lx}

\item{i}{actualisation rate}

\item{dimNames}{parameters other than inc and val (such as generation, sex, age when turning dependent)}

\item{incName}{the x in "lx". Most of the time, age. for dependence tables, usually the nb of years since turning dependent}

\item{valName}{lx}

\item{messages}{= FALSE}

\item{keep_order}{should we keep order of the input table ?}
}
\description{
from lxTable and given act rate, adds commut columns
}
