% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_tables.R
\name{dep_table}
\alias{dep_table}
\title{Long-term care EPV}
\usage{
dep_table(
  inputs,
  detailedRes = F,
  dim_age_dep_name = "dim_age_dep",
  i = 0,
  timer_messages = F
)
}
\arguments{
\item{inputs}{3 tables : incidence rate / mortality / dependents mortality}

\item{detailedRes}{if FALSE : only returns the table of EPV by age}

\item{dim_age_dep_name}{name of the column giving age when turning dependent}

\item{i}{actualisation rate --> not yet}

\item{timer_messages}{Should we call messages from the timer function ?}
}
\description{
compute Expected Present Value of long-term care annuities, in
 order to help for PRC (risk-increase provision ?)
}
