% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_pop.R
\name{inc_pop}
\alias{inc_pop}
\title{increment a pop Table}
\usage{
inc_pop(
  old_pop = generate_pop(100),
  p_resilRate = 2/100,
  p_subRate = 2/100,
  messages = F
)
}
\arguments{
\item{old_pop}{starting pop table}

\item{p_resilRate}{cancellation rate of people in the table}

\item{p_subRate}{new rows to add - depending on the current table size}

\item{messages}{should the function print messages while running (info only)}
}
\description{
From a pop table with "generate_pop()" format, simulate the new population at N+1
}
