% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLifeTables.R
\name{mergeLifeTables}
\alias{mergeLifeTables}
\title{merge (cross-join ?) life tables}
\usage{
mergeLifeTables(XTbl, YTbl, valPatt = "^qx$", xName, yName)
}
\arguments{
\item{XTbl}{"left" table}

\item{YTbl}{"right" table}

\item{valPatt}{pattern for value columns}

\item{xName}{name of X tabl, used to rename duplicate columns}

\item{yName}{name of Y tabl, used to rename duplicate columns}
}
\description{
given 2 prefixed life tables, merges them on common dims and
makes a cross-join for exclusive dims
}
